project/backend/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pymongo==4.5.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0